feat: create dashboard-view, user-management-view, data-view, profile-view
