+++ .env
@@ -1,0 +1,2 @@
+VUE_APP_AUTH_BASE_URI=https://beamyiradat.adauntukkamu.com
+VUE_APP_USER_BASE_URI=https://bepmyiradat.adauntukkamu.com
