+++ src/data/usersData.ts
@@ -1,0 +1,305 @@
+export const allUsers = [
+  {
+    id: 1,
+    name: 'John Doe',
+    email: 'john.doe@example.com',
+    role: 'Admin',
+    status: 'Active',
+    createdAt: '2024-01-15T10:30:00Z',
+  },
+  {
+    id: 2,
+    name: 'Jane Smith',
+    email: 'jane.smith@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-01-20T14:45:00Z',
+  },
+  {
+    id: 3,
+    name: 'Mike Johnson',
+    email: 'mike.johnson@example.com',
+    role: 'Moderator',
+    status: 'Inactive',
+    createdAt: '2024-01-10T09:15:00Z',
+  },
+  {
+    id: 4,
+    name: 'Sarah Wilson',
+    email: 'sarah.wilson@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-01-25T16:20:00Z',
+  },
+  {
+    id: 5,
+    name: 'David Brown',
+    email: 'david.brown@example.com',
+    role: 'User',
+    status: 'Pending',
+    createdAt: '2024-01-28T11:00:00Z',
+  },
+  {
+    id: 6,
+    name: 'Emily Davis',
+    email: 'emily.davis@example.com',
+    role: 'Admin',
+    status: 'Active',
+    createdAt: '2024-01-12T13:30:00Z',
+  },
+  {
+    id: 7,
+    name: 'Robert Miller',
+    email: 'robert.miller@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-01-18T08:45:00Z',
+  },
+  {
+    id: 8,
+    name: 'Lisa Garcia',
+    email: 'lisa.garcia@example.com',
+    role: 'Moderator',
+    status: 'Active',
+    createdAt: '2024-01-22T15:10:00Z',
+  },
+  {
+    id: 9,
+    name: 'Alex Thompson',
+    email: 'alex.thompson@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-01-30T12:00:00Z',
+  },
+  {
+    id: 10,
+    name: 'Maria Rodriguez',
+    email: 'maria.rodriguez@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-02-01T09:30:00Z',
+  },
+  {
+    id: 11,
+    name: 'James Wilson',
+    email: 'james.wilson@example.com',
+    role: 'Admin',
+    status: 'Active',
+    createdAt: '2024-02-03T14:15:00Z',
+  },
+  {
+    id: 12,
+    name: 'Sophia Chen',
+    email: 'sophia.chen@example.com',
+    role: 'User',
+    status: 'Inactive',
+    createdAt: '2024-02-05T11:45:00Z',
+  },
+  {
+    id: 13,
+    name: 'Michael Lee',
+    email: 'michael.lee@example.com',
+    role: 'Moderator',
+    status: 'Active',
+    createdAt: '2024-02-07T16:20:00Z',
+  },
+  {
+    id: 14,
+    name: 'Emma Taylor',
+    email: 'emma.taylor@example.com',
+    role: 'User',
+    status: 'Pending',
+    createdAt: '2024-02-09T10:00:00Z',
+  },
+  {
+    id: 15,
+    name: 'Daniel Anderson',
+    email: 'daniel.anderson@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-02-11T13:30:00Z',
+  },
+  {
+    id: 16,
+    name: 'Olivia Martinez',
+    email: 'olivia.martinez@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-02-13T08:45:00Z',
+  },
+  {
+    id: 17,
+    name: 'William Johnson',
+    email: 'william.johnson@example.com',
+    role: 'Admin',
+    status: 'Active',
+    createdAt: '2024-02-15T15:10:00Z',
+  },
+  {
+    id: 18,
+    name: 'Ava Brown',
+    email: 'ava.brown@example.com',
+    role: 'User',
+    status: 'Inactive',
+    createdAt: '2024-02-17T12:00:00Z',
+  },
+  {
+    id: 19,
+    name: 'Ethan Davis',
+    email: 'ethan.davis@example.com',
+    role: 'Moderator',
+    status: 'Active',
+    createdAt: '2024-02-19T09:30:00Z',
+  },
+  {
+    id: 20,
+    name: 'Isabella Garcia',
+    email: 'isabella.garcia@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-02-21T14:15:00Z',
+  },
+  {
+    id: 21,
+    name: 'Noah Miller',
+    email: 'noah.miller@example.com',
+    role: 'User',
+    status: 'Pending',
+    createdAt: '2024-02-23T11:45:00Z',
+  },
+  {
+    id: 22,
+    name: 'Mia Wilson',
+    email: 'mia.wilson@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-02-25T16:20:00Z',
+  },
+  {
+    id: 23,
+    name: 'Lucas Taylor',
+    email: 'lucas.taylor@example.com',
+    role: 'Admin',
+    status: 'Active',
+    createdAt: '2024-02-27T10:00:00Z',
+  },
+  {
+    id: 24,
+    name: 'Amelia Anderson',
+    email: 'amelia.anderson@example.com',
+    role: 'User',
+    status: 'Active',
+    createdAt: '2024-03-01T13:30:00Z',
+  },
+  {
+    id: 25,
+    name: 'Mason Martinez',
+    email: 'mason.martinez@example.com',
+    role: 'Moderator',
+    status: 'Inactive',
+    createdAt: '2024-03-03T08:45:00Z',
+  },
+]
+
+export const usersData = {
+  users: [
+    {
+      id: 'DEV_alex_demo_taylor',
+      email: 'dev.alex.demo.taylor@devtest.local',
+      phone: '+15550000005',
+      role_id: 2,
+      name: 'DEV - Alex Demo Taylor',
+    },
+    {
+      id: 'DEV_john_admin_smith',
+      email: 'dev.john.admin.smith@devtest.local',
+      phone: '+15550000001',
+      role_id: 1,
+      name: 'DEV - John Admin EDITED',
+    },
+    {
+      id: 'DEV_sarah_user_johnson',
+      email: 'dev.sarah.user.johnson@devtest.local',
+      phone: '+15550000002',
+      role_id: 2,
+      name: 'DEV - Sarah User Johnson',
+    },
+    {
+      id: 'DEV_mike_cs_wilson',
+      email: 'dev.mike.cs.wilson@devtest.local',
+      phone: '+15550000003',
+      role_id: 3,
+      name: 'DEV - Mike CS Wilson',
+    },
+    {
+      id: 'DEV_emma_test_brown',
+      email: 'dev.emma.test.brown@devtest.local',
+      phone: '+15550000004',
+      role_id: 2,
+      name: 'DEV - Emma Test Brown',
+    },
+    {
+      id: 'DEV +25Aug121711',
+      email: 'dev.fatih@mail.com',
+      phone: '+62001001',
+      role_id: 2,
+      name: 'DEV - Fatih',
+    },
+    {
+      id: 'DEV +25Aug122016',
+      email: 'dev.tiara.admin@mail.com',
+      phone: '+62002002',
+      role_id: 2,
+      name: 'DEV - Tiara',
+    },
+    {
+      id: 'DEV-+25Aug122333',
+      email: 'dev.kintan.cs@mail.com',
+      phone: '+62002003',
+      role_id: 2,
+      name: 'DEV-Kintan',
+    },
+    {
+      id: 'DEV-S+25jan031247',
+      email: 'dev.sutejo.cs@mail.com',
+      phone: '+62002004',
+      role_id: 2,
+      name: 'DEV-Sutejo',
+    },
+    {
+      id: 'DEV-K+25jan031248',
+      email: 'dev.klak.cs@mail.com',
+      phone: '+62002005',
+      role_id: 2,
+      name: 'DEV-Klak',
+    },
+    {
+      id: 'DEV-J+25jan031249',
+      email: 'dev.jared.admin@mail.com',
+      phone: '+62002006',
+      role_id: 2,
+      name: 'DEV-Jared',
+    },
+    {
+      id: 'DEV-K+25jan031308',
+      email: 'dev.klise@mail.com',
+      phone: '+62002007',
+      role_id: 2,
+      name: 'DEV - Klise EDITED',
+    },
+    {
+      id: 'DEV-S+25jan031417',
+      email: 'dev.shin@mail.com',
+      phone: '+62002008',
+      role_id: 2,
+      name: 'DEV-Shin',
+    },
+    {
+      id: '044DEV-Karz',
+      email: 'dev.karz.admin@mail.com',
+      phone: '+62002009',
+      role_id: 1,
+      name: 'DEV-Karzadmin',
+    },
+  ],
+}
