+++ src/views/auth/LoginView.vue
@@ -1,0 +1,78 @@
+<script lang="ts">
+export const description = 'A modern and beautiful login form with enhanced styling.'
+</script>
+
+<script setup lang="ts">
+import LoginForm from '../../components/custom/custom-form/LoginForm.vue'
+</script>
+
+<template>
+  <div
+    class="min-h-screen w-full bg-gradient-to-br from-background via-background to-muted/20 relative overflow-hidden"
+  >
+    <!-- Background decorative elements -->
+    <div class="absolute inset-0 overflow-hidden">
+      <!-- Top left decorative circle -->
+      <div class="absolute -top-40 -left-40 w-80 h-80 bg-primary/5 rounded-full blur-3xl"></div>
+
+      <!-- Bottom right decorative circle -->
+      <div class="absolute -bottom-40 -right-40 w-80 h-80 bg-primary/5 rounded-full blur-3xl"></div>
+
+      <!-- Center decorative element -->
+      <div
+        class="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-96 h-96 bg-primary/3 rounded-full blur-3xl"
+      ></div>
+
+      <!-- Grid pattern overlay -->
+      <div
+        class="absolute inset-0 bg-[linear-gradient(rgba(255,255,255,0.02)_1px,transparent_1px),linear-gradient(90deg,rgba(255,255,255,0.02)_1px,transparent_1px)] bg-[size:50px_50px]"
+      ></div>
+    </div>
+
+    <!-- Main content -->
+    <div class="relative z-10 flex min-h-screen w-full items-center justify-center px-4 py-8">
+      <div class="w-full max-w-md">
+        <!-- Logo/Brand section -->
+        <div class="text-center mb-8">
+          <div
+            class="inline-flex items-center justify-center w-16 h-16 bg-primary/10 rounded-2xl mb-4"
+          >
+            <svg class="w-8 h-8 text-primary" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M13 10V3L4 14h7v7l9-11h-7z"
+              />
+            </svg>
+          </div>
+          <h1 class="text-3xl font-bold text-foreground mb-2">MyIradat</h1>
+          <p class="text-muted-foreground">Your secure platform for data management</p>
+        </div>
+
+        <!-- Login form -->
+        <LoginForm />
+      </div>
+    </div>
+  </div>
+</template>
+
+<style scoped>
+/* Custom scrollbar for webkit browsers */
+::-webkit-scrollbar {
+  width: 6px;
+}
+
+::-webkit-scrollbar-track {
+  background: transparent;
+}
+
+::-webkit-scrollbar-thumb {
+  background: hsl(var(--muted-foreground) / 0.3);
+  border-radius: 3px;
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: hsl(var(--muted-foreground) / 0.5);
+}
+</style>
